Quality of Life changes, Mobile Enhancements, and initialization of Event Snapshot System
