+++ src/PageNotFound/PageNotFound.module.css
@@ -1,0 +1,93 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Open+Sans:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
+@import "/src/Butler_Webfont/stylesheet.css";
+.bg {
+  background: gradient(#24475b, #294357);
+  height: 100vh;
+  width: 100vw;
+}
+
+.backGroundImage {
+  height: 100vh;
+  width: 100vw;
+  position: fixed;
+  left: 0;
+  top: 0;
+  z-index: -1;
+  object-fit: cover;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.mainContent {
+  display: flex;
+  flex-direction: row;
+}
+
+.title {
+  font-family: "Butler";
+  font-size: 4.5em;
+  color: whitesmoke;
+  margin: 20px;
+  padding: 0;
+}
+
+.subtitle {
+  margin: 0;
+  padding: 0;
+  font-weight: 200;
+  font-size: 1.5em;
+  margin: 20px;
+}
+
+.explanation {
+  font-family: "Roboto";
+  font-size: 1.5em;
+  padding: 0;
+  font-weight: 200;
+  margin: 20px;
+}
+
+.returnButton {
+  margin: 20px;
+  padding: 10px;
+  border-radius: 5px;
+  border: 1px solid #24475b;
+  background-color: #24475b;
+  color: white;
+  font-family: "Roboto";
+  font-size: 1.5em;
+  font-weight: 200;
+  cursor: pointer;
+  -webkit-transition-duration: 0.3s;
+  transition-duration: 0.3s;
+  -webkit-transition-property: color, background-color;
+  transition-property: color, background-color;
+}
+.returnButton:hover {
+  background-color: white;
+  color: #24475b;
+}
+
+.logo {
+  font-family: "Butler", serif;
+  font-size: 3em;
+  font-weight: 900;
+  color: #24475b;
+  margin: 0;
+  text-align: center;
+}
+
+#parallax {
+  position: absolute;
+  width: 100%;
+  height: 100vh;
+  z-index: -1;
+}
+
+/*# sourceMappingURL=PageNotFound.module.css.map */
