+++ src/common/constants/globalVariables.scss
@@ -17,6 +17,10 @@ $base-color-hover: #2F5D7A;
 
 //IDEAL SPACINGS
 $content-margin-top-desktop: 5%;
+$content-margin-top-mobile: 10%;
+$content-margin-left-desktop: 10%;
+$content-margin-left-mobile: 5%;
+
 
 
 //BUTTON TRANSITION CRUVES
