--- src/components/Authentication/Login.module.css
@@ -1,137 +1,0 @@
-.pageContainer {
-    display: flex;
-    align-items: stretch;
-    height: 100vh;
-    width: 100vw;
-}
-
-.presentationContainer {
-    background-color: #24475B;
-    display: block;
-    position: relative;
-}
-
-.peopleHavingFun {
-    max-height: 100vh;
-    height: 100%;
-    width: 65vw;
-    border: none;
-    max-width: 100vw;
-    aspect-ratio: auto;
-    object-fit: cover;
-    opacity: .8;
-;
-}
-
-.presentationHeader {
-    font-family: Butler;
-    font-weight: 500;
-    font-size: 50px;
-    color: white;
-    max-width: 950px;
-    position: absolute;
-    bottom: 200px;
-    left: 20px;
-}
-
-.formContainer {
-    background-color: #FEFEFE;
-    flex-grow: 1;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    align-items: center;
-}
-
-.loginWidget {
-    display: block;
-    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-    border-radius: 5px;
-    margin: 0 max(50px, auto);;
-}
-
-.loginLogo {
-    font-family: Butler, "Open Sans";
-    font-weight: 900;
-    font-size: 80px;
-    color: black;
-    margin: 0 auto;
-    display: block;
-    text-decoration: none;
-}
-
-.pageContainer {
-    display: flex;
-    flex-direction: row;
-}
-
-.loginWidgetHeader {
-    padding: 40px;
-    padding-bottom: 10px;
-}
-
-.loginSubText {
-    display: block;
-    font-family: Roboto;
-    color: #3a3a3a;
-    margin-top: -12px;
-    text-align: center;
-    font-size: 15px;
-    text-underline-mode: none;
-    text-decoration: none;
-}
-
-.loginSubSubText {
-    display: block;
-    margin: 0 auto;
-    padding-top: 30px;
-    font-family: Pavanam;
-    text-align: center;
-    font-size: 20px;
-    text-decoration: none;
-    color: black;
-}
-
-.loginWidgetFormContainer {
-    display: block;
-    padding-bottom: 40px;
-    width: auto;
-}
-
-.loginButton {
-    display: block;
-    margin: 0 auto;
-    width: auto;
-}
-
-.toastContainer {
-    margin: 0 auto;
-}
-
-@media screen and (max-width: 900px) {
-    .pageContainer {
-        justify-content: center;
-        flex-direction: column-reverse;
-    }
-
-    .presentationContainer {
-        margin: 0 auto;
-    }
-
-    .presentationHeader {
-        font-size: 24px;
-        bottom: 100px;
-    }
-
-    .peopleHavingFun {
-        width: 100vw;
-        height: 30vh;
-    }
-
-    .presentationHeader {
-        width: 60%;
-        text-align: center;
-        position: center;
-    }
-
-}
