+++ src/components/Dashboard/Dashboard.module.css
@@ -1,0 +1,74 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Open+Sans:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
+@import "/src/Butler_Webfont/stylesheet.css";
+.navBarItems {
+  list-style: none;
+  display: flex;
+  flex-direction: row;
+  width: 90vw;
+  margin: 0;
+  padding: 0;
+}
+
+.navBarItem {
+  font-family: Roboto, sans-serif;
+  font-size: 1.5rem;
+  margin: 0.5em;
+  padding: 10px;
+  border-radius: 10px;
+  white-space: nowrap;
+}
+
+.navBarItemActive {
+  background-color: #24475b;
+  color: white;
+  font-weight: bold;
+}
+
+.navBarItem:hover {
+  background-color: #e8e8e8;
+  cursor: pointer;
+}
+
+.pageContainer {
+  display: flex;
+  flex-direction: column;
+  margin-top: 5%;
+  margin-left: 100px;
+}
+
+@media screen and (max-width: 1200px) {
+  .navBarItem {
+    font-size: 1rem;
+  }
+  .navBarItems {
+    padding: 0;
+  }
+  .navBarItemActive {
+    background-color: #24475b;
+    color: white;
+  }
+  .navBarItem:hover {
+    background: #24475b;
+    cursor: pointer;
+  }
+  .navBar {
+    overflow: hidden;
+    width: 85vw;
+  }
+  .pageContainer {
+    margin-top: 60px;
+    align-items: center;
+  }
+}
+@media screen and (max-width: 700px) {
+  .pageContainer {
+    align-items: flex-start;
+    margin-left: 0;
+  }
+  .navBar {
+    overflow: scroll;
+    width: 100vw;
+  }
+}
+
+/*# sourceMappingURL=Dashboard.module.css.map */
