+++ src/components/Nav/Sidebar/NavButton.module.css
@@ -1,0 +1,39 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Open+Sans:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
+@import "/src/Butler_Webfont/stylesheet.css";
+.navButton {
+  padding: 10px;
+  margin: 5px 0;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  -webkit-box-sizing: border-box; /* Safari/Chrome, other WebKit */
+  -moz-box-sizing: border-box; /* Firefox, other Gecko */
+  box-sizing: border-box; /* Opera/IE 8+ */
+  border: hidden;
+  border-collapse: collapse;
+  cursor: pointer;
+}
+
+.selectedButton {
+  background-color: #24475b;
+}
+.selectedButton .navButtonText {
+  color: white;
+}
+
+.navButton:hover {
+  outline: #c0c0c0 1px solid;
+  box-sizing: border-box;
+}
+
+.navButtonText {
+  color: black;
+  font-family: "Roboto", sans-serif;
+  padding: 0 10px;
+  margin-bottom: 0;
+  font-size: 1rem;
+  white-space: nowrap;
+}
+
+/*# sourceMappingURL=NavButton.module.css.map */
