+++ src/components/Nav/Sidebar/NavButton.module.scss
@@ -7,7 +7,6 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-  justify-content: space-between;
   -webkit-box-sizing: border-box; /* Safari/Chrome, other WebKit */
   -moz-box-sizing: border-box;    /* Firefox, other Gecko */
   box-sizing: border-box;         /* Opera/IE 8+ */
@@ -39,9 +38,7 @@
   margin-bottom: 0;
   font-size: 1rem;
   white-space: nowrap;
-  margin: auto;
 }
 
 .iconContainer{
-  margin: auto;
 }
