+++ src/components/Nav/Sidebar/ToggleNavSize.module.scss
@@ -27,6 +27,7 @@
 @media screen and (max-width: 768px) {
     .navigationToggle {
         left: 10px;
+        top: 10px;
     }
 
     .extended {
